pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tunnelctl"
version = "0.1.0"
description = "Reverse tunnel management service for exposing devices behind CGNAT"
requires-python = ">=3.11"
dependencies = [
    "asyncssh>=2.17",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "jinja2>=3.1",
    "aiosqlite>=0.20",
    "python-telegram-bot>=21",
    "httpx>=0.28",
    "python-dotenv>=1.0",
]

[project.scripts]
tunnelctl = "tunnelctl.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "httpx>=0.28",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tunnelctl"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
